feat: remove config file dep from client.py
